+++ framework/adapters/http/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ['placeholder']
+__all__ = ['HttpAdapter', 'InMemoryHttpAdapter', 'HttpHeader', 'HttpMethod', 'HttpRequest', 'HttpResponse', 'HttpStatus']